refactored for votes for when being read and opened
