+++ src/components/SingleReview.jsx
@@ -1,6 +1,8 @@
 import { useParams } from "react-router-dom";
 import { useEffect, useState } from "react";
 
+import Votes from "./Votes";
+
 const SingleReview = () => {
   const [isLoading, setIsLoading] = useState(true);
   const [review, setReviews] = useState([]);
@@ -32,8 +34,8 @@ const SingleReview = () => {
       <p>Comment count: {review.comment_count}</p>
       <p>Comments: {review.review_body}</p>
       <p>Posted by {review.owner} </p>
-      <p>Comments: {review.votes}</p>
       <p>{review.created_at}</p>
+      <Votes review={review} />
     </div>
   );
 };
